Improve PartitionedTopicsSchemaTest
*Modifications*

- Add `SubscriptionInitialPosition` when constructing consumers to provide a predictable consumption
  behavior for testings
- Create a test base for bk ensemble tests `BkEnsemblesTestBase`
- Make PartitionedTopicsSchemaTest extend BkEnsemblesTestBase
